Refactor druid.go into focused files for better maintainability   * client.go - Core Client struct and main methods   * request.go - Request building with extracted helper methods   * retry.go - Retry and error handling logic   * options.go - Client options and configuration   * response.go - Response handling and errors
